+++ src/auto_smart_factory/src/agent/path_planning/Map.cpp
@@ -96,7 +96,7 @@ TimedLineOfSightResult Map::whenIsTimedLineOfSightFree(const Point& pos1, double
 	return result;
 }
 
-bool Map::isTimedConnectionFree(const Point& pos1, const Point& pos2, double startTime, float waitingTime, float drivingTime) const {
+bool Map::isTimedConnectionFree(const Point& pos1, const Point& pos2, double startTime, double waitingTime, double drivingTime) const {
 	// Does not check against static obstacles, this is only used to verify a already planned connection
 	
 	double endTime = startTime + waitingTime + drivingTime;
